requirements.txt
-i https://pypi.org/simple
#glcontext==3.0.0
#memory-profiler==0.61.0; python_version >= '3.5'
moderngl==5.12.0; python_version >= '3.7'
pillow==11.2.1; python_version >= '3.9'
#psutil==7.0.0; python_version >= '3.6'
pygame==2.6.1; python_version >= '3.6'
pyglm==2.8.2
#pytmx==3.32; python_version >= '3.7'
setuptools==80.9.0; python_version >= '3.9'
